+++ src/pages/Login/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 // import ProconLogo from "@/assets/procon-logo.svg";
 import { ProconLogo } from "@/assets/procon-logo";
-import { FooterSupport } from "@/components/footer-support";
+import { FooterSupport } from "@/components/footerSupport";
 import { LoginForm } from "@/components/forms/Login";
 
 export const Login: React.FC = () => {
